Add additional runner.sh steps for M1 macs
Homebrew installs to a different location on M1 machines than on Intel.
This new path is not on the PATH by default and requires modifying the
user's shell configuration to add.

These commands were copy-pasted from the Homebrew install script's
output, just doubled up to make sure that both zsh and bash user's were
accounted for.

Some (maybe most?) casks also require Rosetta to be installed (such as
the JVM); this is also added to the configuration.

This commit roughly mirrors the commands I ran manually to get my
machine ready to run the remaining parts of the script.