Add additional runner.sh steps for M1 macs (#5)
* Add additional runner.sh steps for M1 macs

Homebrew installs to a different location on M1 machines than on Intel.
This new path is not on the PATH by default and requires modifying the
user's shell configuration to add.

These commands were copy-pasted from the Homebrew install script's
output, just doubled up to make sure that both zsh and bash user's were
accounted for.

Some (maybe most?) casks also require Rosetta to be installed (such as
the JVM); this is also added to the configuration.

This commit roughly mirrors the commands I ran manually to get my
machine ready to run the remaining parts of the script.

* Separate M1 detection logic from homebrew

We should install Rosetta by default on M1 systems because so many
programs aren't built for ARM.

* Improve output from runner.sh for M1 macs

Only echo information about Rosetta's installation if it needs to be
installed.

Fix redirect of `command -v` to never print anything. Previously it
redirected stderr, which will always be empty, and left the path of the
command printed to the shell. Nothing will be shown to the user now